ai_pdf_agent/requirements.txt
flask
pdfplumber
spacy
openai
python-dotenv
streamlit
